% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RdataCrypt.R
\name{EncryptRdata}
\alias{EncryptRdata}
\title{Encrypt Rdata files}
\usage{
EncryptRdata(filename = ".Rdata", path = getwd(), RmOld = T, askPW = T,
  ...)
}
\arguments{
\item{filename}{The name of the Rdata file.}

\item{path}{The directory where the Rdata file is stored. Default is current working directory.}

\item{RmOld}{Logical, indicates whether the Rdata file has to be removed.}

\item{askPW}{Logical, indicates whether the password has to be given using the \code{\link[rstudioapi]{askForPassword}} function.}

\item{...}{Arguments to be passed to \code{\link{encrypt}}}
}
\value{
The Rdata file is encrypted and will be saved in the same directory as specified in the \code{path} argument.
}
\description{
Encrypt Rdata files
}
