% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RdataCrypt.R
\name{LoadRcrypt}
\alias{LoadRcrypt}
\title{Load encrypted Rdata files}
\usage{
LoadRcrypt(filename = ".Rdata.gpg", path = getwd(), RmOrig = T, ...)
}
\arguments{
\item{filename}{Name of the encrypted file. E.g. \code{"Test.Rdata.gpg"}.}

\item{path}{Directory where the file is located. Default is current working directory.}

\item{RmOrig}{Logical, indicates whether the original file should be removed. Default is \code{TRUE}.}

\item{...}{Arguments to be passed to \code{\link{decrypt}}.}
}
\value{
The encrypted Rdata file is loaded into the global environment.
}
\description{
Load encrypted Rdata files
}
