% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenRAR.R
\name{OpenRAR}
\alias{OpenRAR}
\title{Open encrypted .RAR files}
\usage{
OpenRAR(pathRAR, pathRARexe = "C:/Program Files/Winrar",
  pathExtract = paste0(getwd(), "/tmp"), pw = c(T, F), RmFiles = T,
  RmTime = "00:30:00", pathRscript = getwd())
}
\arguments{
\item{pathRAR}{The path to the encrypted .rar file. For example, \code{"C:/Users/u007/Documents/MyNameisRAR_WinRAR.rar"}.}

\item{pathRARexe}{Path where the WinRAR executable is located. Default is \code{"C:/Program Files/Winrar"}.}

\item{pathExtract}{Path where the files have to be extracted to. This folder and all its files will be deleted if 
\code{RmFiles = T}.}

\item{pw}{Logical, indicates whether the .rar file is password protected. If \code{TRUE}, the user will be prompted to give the
password.}

\item{RmFiles}{Logical, indicates whether the extracted files have to be deleted.}

\item{RmTime}{The time until removal of the extracted files.}

\item{pathRscript}{The path where the temporary Rscript has to be saved.}
}
\value{
The extracted files will be saved in pathExtract folder and if specified, removed after a certain time.
}
\description{
Open encrypted .RAR files
}
\details{
The extracted files and temporary folder will automatically deleted after the specified time period IF AND
ONLY IF the windows command window is not closed. In addition, there will be some additional temporary files left
that were created to execute the command (an R-script and a .bat file). These can be removed by using the function
RmRemainingFiles.
}
