\name{RcryptAdj-package}
\alias{RcryptAdj}
\docType{package}
\title{
  \packageTitle{RcryptAdj}
}
\description{
  This package is partially based on the \code{rcrypt} package. The \code{encrypt} function has been adjusted to be able
  to encrypt files with different keys and the \code{decrypt} function was not adjusted. Further, additional functionality has been
  added by adding new functions. Users can now quickly encrypt .Rdata files, open encrypted .RAR files, as well as write and open
  (password protected) excel files.
}
\author{
  BavoDC
  Maintainer: BavoDC
}
\keyword{ package }

