% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExcelFntions.R
\name{read_excel_allsheets_pw}
\alias{read_excel_allsheets_pw}
\title{Function to read all sheets of a password protected excel file}
\usage{
read_excel_allsheets_pw(path, pw = NULL, rmHist = T)
}
\arguments{
\item{path}{The path to the excel file.}

\item{pw}{Password of the excel file. Default is \code{NULL} and the user is then prompted to
give the password using \code{\link[rstudioapi]{askForPassword}}.}

\item{rmHist}{Logical, indicates whether the history has to be deleted.}
}
\value{
Returns a list with all the excel sheets.
}
\description{
Function to read all sheets of a password protected excel file
}
\details{
This function needs the package \code{RDCOMClient}, which can be installed using: \cr
\code{library(devtools)} \cr
\code{install_github('omegahat/RDCOMClient')} \cr
}
