% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExcelFntions.R
\name{read_excel_pw}
\alias{read_excel_pw}
\title{Import password protected excel files}
\usage{
read_excel_pw(path, sheet = NULL, pw = NULL, rmHist = T)
}
\arguments{
\item{path}{The path to the excel file.}

\item{sheet}{The sheet that has to be read. Default is \code{NULL}, which causes the first sheet to be imported.}

\item{pw}{Password of the excel file. Default is \code{NULL} and the user is then prompted to
give the password using \code{\link[rstudioapi]{askForPassword}}.}

\item{rmHist}{Logical, indicates whether the history has to be deleted.}
}
\value{
Returns the excel file as a dataframe.
}
\description{
Import password protected excel files
}
\details{
This function needs the package \code{RDCOMClient}, which can be installed using: \cr
\code{library(devtools)} \cr
\code{install_github('omegahat/RDCOMClient')} \cr
}
