% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExcelFntions.R
\name{savexlsx}
\alias{savexlsx}
\title{Write a dataframe or list to an excel file (with or without password).}
\usage{
savexlsx(Object, path = paste0(getwd(), "/", deparse(substitute(Object)),
  ".xlsx"), SheetName = NULL, pw = F, rmHist = pw)
}
\arguments{
\item{Object}{Input object, must be of type dataframe or list.}

\item{path}{The path for the excel file. Default is current working directory.}

\item{SheetName}{Optional, name of the sheet.}

\item{pw}{Logical, indicates whether the file has to be password protected or not. If \code{TRUE}, the user will be prompted to
give the password using \code{\link[rstudioapi]{askForPassword}}.}
}
\description{
Write a dataframe or list to an excel file (with or without password).
}
\examples{

# Without missing values
data(mtcars)
savexlsx(mtcars)

# With missing values
data('mtcars')
mtcars[1, ] = NA
test = xlsxFix(mtcars)
savexlsx(test)

}
\seealso{
\code{\link{xlsxFix}}
}
