% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExcelFntions.R
\name{xlsxFix}
\alias{xlsxFix}
\title{Temporary fix for savexlsx function if there are missing values.}
\usage{
xlsxFix(x)
}
\arguments{
\item{x}{An object of class dataframe or a list.}
}
\value{

}
\description{
Temporary fix for savexlsx function if there are missing values.
}
\details{
This function has to be used in case of missing values. If not, the missing values will be replaced by large values
in the excel file.
}
\examples{
data('mtcars')
mtcars[1, ] = NA
test = xlsxFix(mtcars)
savexlsx(test)

}
